requirements.txt
awsebcli==3.14.6
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==3.0.2
blessed==1.15.0
boto3==1.9.50
botocore==1.12.50
cached-property==1.5.1
cement==2.8.2
certifi==2018.10.15
chardet==3.0.4
colorama==0.3.9
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-extensions==2.1.3
django-storages==1.7.1
djangorestframework==3.9.0
docker==3.5.1
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
idna==2.6
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
lxml==4.2.5
MarkupSafe==1.1.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.2
oauthlib==2.1.0
pandocfilters==1.4.2
parso==0.3.1
pathspec==0.5.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.18.4
requests-oauthlib==1.0.0
s3transfer==0.1.13
semantic-version==2.5.0
Send2Trash==1.5.0
sentry-sdk==0.5.5
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
sqlparse==0.2.4
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
texttable==0.9.1
tornado==5.1.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0
widgetsnbextension==3.4.2
